requirements.txt
Pillow==10.4.0
scikit-image==0.24.0
opencv-python==4.10.0.84
numpy==2.0.0
